add: twig ems_version function
