+++ src/Resources/config/contracts.xml
@@ -7,8 +7,6 @@
     <services>
         <defaults public="false" />
 
-        <service id="EMS\CommonBundle\Contracts\Composer\ComposerInfoInterface" alias="ems_common.composer.info" />
-
         <service id="EMS\CommonBundle\Contracts\CoreApi\CoreApiFactoryInterface" alias="ems_common.core_api.factory" />
         <service id="EMS\CommonBundle\Contracts\Elasticsearch\QueryLoggerInterface" alias="ems_common.elasticsearch.elastica_logger" />
