+++ src/Twig/CommonExtension.php
@@ -21,6 +21,7 @@ class CommonExtension extends AbstractExtension
             new TwigFunction('ems_html', [TextRuntime::class, 'emsHtml'], ['is_safe' => ['all']]),
             new TwigFunction('ems_nested_search', [SearchRuntime::class, 'nestedSearch']),
             new TwigFunction('ems_image_info', [AssetRuntime::class, 'imageInfo']),
+            new TwigFunction('ems_version', [InfoRuntime::class, 'version']),
         ];
     }
